+++ Ascii24D.sln
@@ -10,8 +10,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Engine", "Engine\Engine.csp
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ConsoleWindows", "ConsoleWindows\ConsoleWindows.vcxproj", "{CC03F25A-CB08-472E-A1E6-2A3152EBA58A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EngineTests", "EngineTests\EngineTests.csproj", "{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demos", "Demos", "{219AFD4F-A94D-4A8D-AAB2-2FE198501651}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DirectInputDebugDemo", "Demos\DirectInputDebugDemo\DirectInputDebugDemo.csproj", "{17302DB7-7439-43E9-AAE6-9B3B0FD1B194}"
@@ -22,6 +20,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Editor", "Editor\Editor.csproj", "{159441AC-5BED-4071-A415-22D9E009F529}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Peer2PeerNetworkingDemo", "Demos\Peer2PeerNetworkingDemo\Peer2PeerNetworkingDemo.csproj", "{8841A436-A55F-4679-8745-DC7CB8853B44}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EngineIntegrationTests", "EngineIntegrationTests\EngineIntegrationTests.csproj", "{8628F198-9E61-45DB-814D-55C5BCC5C23B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EngineUnitTests", "EngineUnitTests\EngineUnitTests.csproj", "{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,18 +59,6 @@ Global
 		{CC03F25A-CB08-472E-A1E6-2A3152EBA58A}.Release|x64.Build.0 = Release|x64
 		{CC03F25A-CB08-472E-A1E6-2A3152EBA58A}.Release|x86.ActiveCfg = Release|Win32
 		{CC03F25A-CB08-472E-A1E6-2A3152EBA58A}.Release|x86.Build.0 = Release|Win32
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|x64.Build.0 = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Debug|x86.Build.0 = Debug|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|x64.ActiveCfg = Release|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|x64.Build.0 = Release|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|x86.ActiveCfg = Release|Any CPU
-		{B5EF9D72-2567-4ED4-B7F5-EC8304BDF3F2}.Release|x86.Build.0 = Release|Any CPU
 		{17302DB7-7439-43E9-AAE6-9B3B0FD1B194}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{17302DB7-7439-43E9-AAE6-9B3B0FD1B194}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{17302DB7-7439-43E9-AAE6-9B3B0FD1B194}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -105,6 +95,30 @@ Global
 		{8841A436-A55F-4679-8745-DC7CB8853B44}.Release|x64.Build.0 = Release|Any CPU
 		{8841A436-A55F-4679-8745-DC7CB8853B44}.Release|x86.ActiveCfg = Release|Any CPU
 		{8841A436-A55F-4679-8745-DC7CB8853B44}.Release|x86.Build.0 = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|x64.Build.0 = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Debug|x86.Build.0 = Debug|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|x64.ActiveCfg = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|x64.Build.0 = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|x86.ActiveCfg = Release|Any CPU
+		{8628F198-9E61-45DB-814D-55C5BCC5C23B}.Release|x86.Build.0 = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|x64.Build.0 = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Debug|x86.Build.0 = Debug|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|x64.ActiveCfg = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|x64.Build.0 = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|x86.ActiveCfg = Release|Any CPU
+		{4B51ACDE-BB5E-4925-98D4-70B6F878BA67}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
